Refactor command handling
This is much cleaner and easier to add to,
MatrixBotModule.add_commands() is inspired by argparse and allows easy
specification of the command name(s), its arguments, validators for
the arguments, and the callback function.

This also allows auto generation of the help. "user_" is included for
potential future event handling when triggered by private messages, or
indeed to pass along the sender, even tho this can be derived from
"event".