Send a dummy char to get out of follow-up prompts
If a verb could be applied to multiple objects, then the
game (currently make it good, but others as well) asks for
specification in a second prompt, which looks the same as the main
prompt, but it won't accept "save", so the follow up saving breaks.

The '#\n' breaks out of such a prompt and also is harmless when
applied after a command that ends up in the normal prompt, so for now
this seems like a sufficient workaround.